Create first try FIFO LIFO.smbp
trying LIFO FIFO FUNCTION BLOCKS. there are 3 inputs on this block, which are reset, store and retrieve. when we want to store the value to the block we have to give store input high. there are 2 outputs, free and full. free indicates that register is empty whilst full indicates register is full